% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corona19-package.R
\docType{package}
\name{corona19-package}
\alias{corona19}
\alias{corona19-package}
\title{corona19: Data Package of Coronavirus Disease 2019 in South Korea}
\description{
Get datasets of Coronavirus disease 2019 in South Korea.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/youngwoos/corona19}
  \item Report bugs at \url{https://github.com/youngwoos/corona19/issues}
}

}
\author{
\strong{Maintainer}: Youngwoo Kim \email{stats7445@gmail.com}

}
\keyword{internal}
