% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_returns.R
\name{log_returns}
\alias{log_returns}
\title{Daily Log Returns}
\usage{
log_returns(asset, price1, price2)
}
\arguments{
\item{asset}{A list containing prices, dividends and splits (if any)}

\item{price1}{Price of the FIRST day in time for which you wish to calculate log returns}

\item{price2}{Price of the FINAL day in time for which you wish to calculate log returns}
}
\description{
Get log returns between two dates for a dividend-paying asset
}
