% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corona19-package.R
\docType{package}
\name{Corona19-package}
\alias{Corona19}
\alias{Corona19-package}
\title{Corona19}
\description{
Explore Coronavirus Data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakevestal.github.io/Corona19/}
  \item Report bugs at \url{https://jakevestal.github.io/Corona19/issues}
}

}
\author{
\strong{Maintainer}: Jake Vestal \email{jmvestal@gmail.com} (\href{https://orcid.org/0000-0002-2970-4795}{ORCID})

}
\keyword{internal}
